# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/invoke_info/status": {
            "post": {
                "description": "Get invoke information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoke"
                ],
                "summary": "Get invoke information",
                "parameters": [
                    {
                        "description": "invoke request",
                        "name": "invokeRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetInvokeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
        "/api/invoke": {
            "post": {
                "description": "Invoke interface",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoke"
                ],
                "summary": "Invoke interface",
                "parameters": [
                    {
                        "description": "invoke request",
                        "name": "invokeRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InvokeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "middlewares.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.GetInvokeRequest": {
            "type": "object",
            "required": [
                "method",
                "path"
            ],
            "properties": {
                "method": {
                    "description": "请求类型",
                    "type": "string",
                    "enum": [
                        "GET",
                        "POST",
                        "PUT",
                        "DELETE"
                    ]
                },
                "path": {
                    "description": "接口地址",
                    "type": "string"
                }
            }
        },
...
        "models.InvokeRequest": {
            "type": "object",
            "required": [
                "method",
                "url"
            ],
            "properties": {
                "method": {
                    "description": "请求类型",
                    "type": "string",
                    "enum": [
                        "GET",
                        "POST",
                        "PUT",
                        "DELETE"
                    ]
                },
                "url": {
                    "description": "接口地址",
                    "type": "string"
                }
            }
        },
...
    }
}